Mueve codigo de array a otra clase
